Re #INTRD-27985 Fix Custom table record deletion - change from BigInteger to Long that now Hibernate 6+ returns

+++ opencell-admin/ejbs/src/main/java/org/meveo/service/base/NativePersistenceService.java
@@ -1683,13 +1683,13 @@ public class NativePersistenceService extends BaseService {
     }
 
     @SuppressWarnings("unchecked")
-    public List<BigInteger> filterExistingRecordsOnTable(String tableName, List<Long> ids) {
+    public List<Long> filterExistingRecordsOnTable(String tableName, List<Long> ids) {
         tableName = addCurrentSchema(tableName);
         Session session = getEntityManager().unwrap(Session.class);
         StringBuilder selectQuery = new StringBuilder("select ").append(FIELD_ID).append(" from ").append(tableName).append(" e where ").append(FIELD_ID).append(" in (:ids)");
         NativeQuery query = session.createNativeQuery(selectQuery.toString());
         query.setParameterList("ids", ids);
-        return (List<BigInteger>) query.setFlushMode(FlushModeType.COMMIT).list();
+        return (List<Long>) query.setFlushMode(FlushModeType.COMMIT).list();
     }
 
     /**
